==> 9 Kaggle/1.1.ipynb <==
Data sourced from https://www.kaggle.com/datasets/bharatnatrayn/movies-dataset-for-feature-extracion-prediction/

### How can we detect problems with the data?
- Sample the data using `.head()` and look for problems
- Count `NaN` values
- Use `.value_counts()` to find any non-NaN default values

import pandas

df = pandas.read_csv("./1.0.csv")
df.head()

print("% of non-null values")
df.count() / df.shape[0]

for column in df.columns:
    print(f"---- {column} ---")
    print(df[column].value_counts().head(5))

| What problems should we worry about?                                                | What can we do about these problems?                                                                                                                                                            |
| ----------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
| Useless index                                                                       | Make the MOVIES field the index                                                                                                                                                                 |
| Many duplicate values in `MOVIES`                                                   | Drop all rows that are completely identical, compare resulting duplicates                                                                                                                       |
| Years Running are stored in one string field                                        | Parse with RegEx <br>- split into a start and end field <br>- leave end as NaN if still running <br>- leave both as NaN if source is NaN <br>- correctly handle shows that only lasted one year |
| `GENRE` field is oddly formatted                                                    | Trim & split on `, `                                                                                                                                                                            |
| NaN values in `VOTES`                                                               | Default `VOTES` to `0`                                                                                                                                                                          |
| Extraneous whitespace in `ONE-LINE`                                                 | Trim field                                                                                                                                                                                      |
| `\nAdd a Plot\n` as a default value for `ONE-LINE`                                  | Replace with NaN                                                                                                                                                                                |
| Confusing field name for `ONE-LINE`                                                 | Manually rename to `decription`                                                                                                                                                                 |
| Many fields combined into the `STARS` field                                         | Parse with string operators, split into `directors` and `stars` fields                                                                                                                          |
| `Gross` field is mostly NaN                                                         | Drop column as it wouldn't be useful enough for anything                                                                                                                                        |
| Inconsistant capitalization in field names (`SCREAMING-KEBAB-CASE` vs `PascalCase`) | Manually rename all to use `snake_case`                                                                                                                                                         |
