==> 9 Kaggle/1.4.ipynb <==
import pandas

df = pandas.read_csv("./1.3.csv", index_col="name")  # Monday's changes included
df

# Drop all rows that are completely identical
# Ignore the `description` column, we can't easily resolve those conflicts so only keep the first
df.drop_duplicates(
    inplace=True,
    keep="first",
    subset=["years", "genres", "rating", "stars", "votes", "run_time", "directors"],
)

# Count still-existing `name` duplicates
dupe_counts = df.index.value_counts()
print(dupe_counts[dupe_counts > 1].head(20))
print(sum(dupe_counts[dupe_counts > 1]))
# 60% of the data is duplicated what is this

df[df.index.duplicated(keep=False)].sort_index().head(20)

# Looks like many individual TV show episodes are included
# We can just leave those as-is
# We should bring back the numerical index if the name isn't going to be unique then
df.reset_index(inplace=True)

df["dupe_counts"] = df["name"].map(lambda name: dupe_counts[name])
duplicates = df[df["dupe_counts"] > 1]
duplicates.sort_values(["dupe_counts", "name"]).head(20)

If the description is NaN, there's no way to distinguish different episodes
Average out run_time, votes, rating
Concat genres, stars, directors
Use the older year started and newer year ended (do it after `years` is parsed)

Parse `years` with RegEx 
- split into a start and end field
- leave end as NaN if still running
- leave both as NaN if source is NaN
- correctly handle shows that only lasted one year

df.to_csv("./1.4.csv", index=False)
