==> 9 Kaggle/1.5.ipynb <==
import pandas

pandas.set_option("display.max_rows", 100)
df = pandas.read_csv("./1.4.csv")  # Tuesday's changes included
df

Parse `years` with RegEx 
- split into a start and end field
- leave end as NaN if still running
- leave both as NaN if source is NaN
- correctly handle shows that only lasted one year

years = df["years"].str.extract(r"\((?P<start>\d{4})(?: \w|(?:–(?P<end> |\d{4}))?\)$)")
years["source"] = df["years"]
years

# Any remaining unparsable sources?
years[~years["start"].astype(str).str.isdecimal()]["source"].value_counts()
# After my first attempt, I hadn't accounted for values that includes roman numerals or other note AS WELL as a year
# Now, the only unparasble values don't have years in them
# Note that .value_counts() doesn't count `NaN` values in the `source` column

# Replace values in the `end` column
# If it's `NaN`, it only lasted a year. Set to `start`
# If it's `" "`, it's still running. Set to `None`
from numpy import NaN

years["end"].fillna(years["start"], inplace=True)
years["end"].replace(" ", NaN, inplace=True)
years

df["start_year"] = years["start"]
df["end_year"] = years["end"]
df.drop(columns=["years", "dupe_counts"], inplace=True)

Will finish tomorrow

df.to_csv("./1.5.csv", index=False)
