==> 9 Kaggle/2.1.ipynb <==
Data sourced from https://www.kaggle.com/datasets/rashikrahmanpritom/data-science-job-posting-on-glassdoor/

### How can we detect problems with the data?

-   Sample the data using `.head()` and look for problems
-   Use `.dtype` to verify datatypes
-   Use `.value_counts()` to find any non-NaN default values or common preceeding whitespace in values
-   Use `.duplicated()` to check that there are no duplicate values


import pandas

df = pandas.read_csv("./2.0.csv")
df.head()

df.dtypes

df.size

for column in df.columns:
    print(f"---- {column} ----")
    print(df[column].value_counts().head(5))
    if df[column].dtype == object:
        print("-- First characters --")
        print(df[column].str[0].value_counts().head(5))

df[df.duplicated()]

| What problems should we worry about?                 | What can we do about these problems?                                     |
| ---------------------------------------------------- | ------------------------------------------------------------------------ |
| Extra index column in the source dataset             | Drop it                                                                  |
| `Salary Estimate`, `Size`, and `Revenue` are strings | Convert to tuples of numbers                                             |
| `-1` as a default value nearly everywhere            | Replace all with `NaN`                                                   |
| `Rating` is included in the `Company Name`           | Parse with string operators                                              |
| `Competitors` are stored as a string                 | Split and replace with an array                                          |
| All field values are in _Title Case_                 | Rename all to use _snake_case_ so they're easier to use programmatically |
