==> 9 Kaggle/2.2.ipynb <==
| What problems should we worry about?                   | What can we do about these problems?                                     |
| ------------------------------------------------------ | ------------------------------------------------------------------------ |
| Extra index column in the source dataset               | Drop it                                                                  |
| `Salary Estimate`, `Size`, and `Revenue` are strings   | Convert to tuples of numbers                                             |
| `-1` as a default value nearly everywhere              | Replace all with `NaN`                                                   |
| `Rating` is included in the `Company Name`             | Parse with string operators                                              |
| `Competitors` are stored as a string                   | Split and replace with an array                                          |
| All field values are in _Title Case_                   | Rename all to use _snake_case_ so they're easier to use programmatically |
| `Unknown / Non-Applicable` default value for `Revenue` | Replace with `NaN`                                                       |


import pandas

df = pandas.read_csv("./2.0.csv")
df.head()

df.drop(columns="index", inplace=True)
# Drp extra index column

df.rename(
    columns={
        "Job Title": "job_title",
        "Salary Estimate": "salary",
        "Job Description": "description",
        "Rating": "rating",
        "Company Name": "company",
        "Headquarters": "headquarters",
        "Size": "size",
        "Founded": "founded",
        "Type of ownership": "ownership",
        "Industry": "industry",
        "Sector": "sector",
        "Revenue": "revenue",
        "Competitors": "competitors",
    },
    inplace=True,
)
# snake_case all columns

df

from numpy import NaN

df.replace(-1, NaN, inplace=True)  # Replace values in rating, founded, etc
df.replace("-1", NaN, inplace=True)  # Replace values in sector, competitors, etc
df["revenue"].replace(
    "Unknown / Non-Applicable", NaN, inplace=True
)  # Replace values in revenue

print("% of non-null values")
df.count() / df.shape[0]
# Remeasure

# Drop competitors column as a large portion of it is NaN, and it's not directly related to the job posting either
df.drop(columns="competitors", inplace=True)

| What problems should we worry about?                 | What can we do about these problems?                                     |
| ---------------------------------------------------- | ------------------------------------------------------------------------ |
| `Salary Estimate`, `Size`, and `Revenue` are strings | Convert to tuples of numbers                                             |
| `Rating` is included in the `Company Name`           | Parse with string operators                                              |
| `Competitors` are stored as a string                 | Split and replace with an array                                          |


df.to_csv("./2.2.csv", index=False)
