==> 9 Kaggle/2.3.ipynb <==
| What problems should we worry about?                 | What can we do about these problems?                                     |
| ---------------------------------------------------- | ------------------------------------------------------------------------ |
| `Salary Estimate`, `Size`, and `Revenue` are strings | Convert to tuples of numbers                                             |
| `Rating` is included in the `Company Name`           | Parse with string operators                                              |


import pandas

df = pandas.read_csv("./2.2.csv")
df.head()

df["company"] = df["company"].apply(lambda info: info.split("\n")[0])
df  # Remove rating data from the company name

df

df["size"].value_counts()

from numpy import NaN

df["size"].replace("Unknown", NaN, inplace=True)

def handle_size(size):
    if not isinstance(size, str):
        return size
    sections = size.split()
    return (sections[0], sections[2]) if len(sections) > 2 else (10000,)


df["size"] = df["size"].apply(handle_size)

| What problems should we worry about?                 | What can we do about these problems?                                     |
| ---------------------------------------------------- | ------------------------------------------------------------------------ |
| `Salary Estimate` and `Revenue` are strings | Convert to tuples of numbers                                             |

df.to_csv("./2.3.csv", index=False)
