==> 9 Kaggle/3.2.ipynb <==
import pandas

df = pandas.read_csv("./3.0.csv")
df.head()

| What problems should we worry about?                         | What can we do about these problems?   |
| ------------------------------------------------------------ | -------------------------------------- |
| All field values are in Spanish and lowercase without spaces | Rename all to use English _snake_case_ |
| The order of columns is unintuitive                          | Sort them                              |
| `idInformation` should be the index                          | Set it to the index                    |
| `periodoreportafin` and `periodoreportainicio` aren't dates  | Convert them                           |
| Many unavailable `nombre` values should be normalized        | Replace                                |
| `"Ver nota"` and `"na"` as default values for `denominacion` | Replace with `NaN`                     |
| `cargo` and `denominacion` appear to be identical            | Drop one                               |


df[
    (df["denominacion"] != df["cargo"])
    & (df["denominacion"].notna() | df["cargo"].notna())
]

df.drop(columns="denominacion", inplace=True)

df.rename(
    columns={
        "nombre": "name",
        "cargo": "position",
        "area": "department",
        "sujetoobligado": "authority",
        "entidadfederativa": "state",
        "montoneto": "net",
        "montobruto": "gross",
        "periodoreportainicio": "reporting_start",
        "periodoreportafin": "reporting_end",
        "idInformacion": "id_information",
    },
    inplace=True,
)
# English snake_case all columns

df = df[[
    "id_information",
    "name",
    "position",
    "department",
    "authority",
    "state",
    "net",
    "gross",
    "reporting_start",
    "reporting_end",
]]
df

df.set_index("id_information")

print("% of non-null values")
df.count() / df.shape[0]

df.to_csv("./3.2.csv", index=False)

| What problems should we worry about?                         | What can we do about these problems?   |
| ------------------------------------------------------------ | -------------------------------------- |
| `periodoreportafin` and `periodoreportainicio` aren't dates  | Convert them                           |
| Many unavailable `nombre` values should be normalized        | Replace                                |
| `"Ver nota"` and `"na"` as default values for `denominacion` | Replace with `NaN`                     |
