==> 9 Kaggle/3.3.ipynb <==
import pandas

df = pandas.read_csv("./3.2.csv", index_col="id_information")
df.head()

| What problems should we worry about?                         | What can we do about these problems?   |
| ------------------------------------------------------------ | -------------------------------------- |
| `periodoreportafin` and `periodoreportainicio` aren't dates  | Convert them                           |
| Many unavailable `nombre` values should be normalized        | Replace                                |
| `"Ver nota"` and `"na"` as default values for `denominacion` | Replace with `NaN`                     |


df.index.rename("id", inplace=True)

def parse(column: str):
    source = df[column].replace("/0208", "/2018", regex=True)
    one = pandas.to_datetime(source, format="%d/%m/%Y", errors="coerce")
    two = pandas.to_datetime(source, format="%d/%m/00%y", errors="coerce")
    three = pandas.to_datetime(source, format="%d/%m/02%y", errors="coerce")
    df[column] = one.fillna(two).fillna(three)


parse("reporting_start")
parse("reporting_end")

df.head()

df[df["reporting_start"].isna() | df["reporting_end"].isna()]

df.dtypes

df.to_csv("./3.3.csv", index=False)
