==> 9 Kaggle/3.4.ipynb <==
import pandas

df = pandas.read_csv(
    "./3.3.csv", index_col="id", parse_dates=["reporting_start", "reporting_end"]
)
df.head()

| What problems should we worry about?                     | What can we do about these problems? |
| -------------------------------------------------------- | ------------------------------------ |
| Many unavailable `name` values should be normalized      | Find and replace them                |
| `"Ver nota"` and `"na"` as default values for `position` | Replace with `NaN`                   |


from numpy import NaN

df.position.replace("na", NaN, inplace=True)
df.position.replace("Ver nota", NaN, inplace=True)
df.position.value_counts()[:20]

df = df.apply(
    lambda series: (
        series.str.strip().replace(r"\s\s+", " ", regex=True)
        if series.dtype == object
        else series
    )
)

df.to_csv("./3.4.csv")
